DEV-30 Refactored main and dataTransfer
