+++ src/Shared/streaming_handlers/web_adapters.go
@@ -1,9 +1,9 @@
 package streaming
 
 import (
-	"github.com/HyperloopUPV-H8/Backend-H8/DataTransfer/domain"
-	"github.com/HyperloopUPV-H8/Backend-H8/DataTransfer/domain/measurement"
 	packetParser "github.com/HyperloopUPV-H8/Backend-H8/Shared/packet_adapter/packet_parser/domain"
+	"github.com/HyperloopUPV-H8/Backend-H8/data_transfer/domain"
+	"github.com/HyperloopUPV-H8/Backend-H8/data_transfer/domain/measurement"
 )
 
 type PacketWebAdapter struct {
