+++ src/cmd/MVP-2/main.go
@@ -5,14 +5,17 @@ import (
 	"os"
 	"time"
 
-	dataTransfer "github.com/HyperloopUPV-H8/Backend-H8/DataTransfer"
-	logger "github.com/HyperloopUPV-H8/Backend-H8/Shared/Logger/infra"
 	server "github.com/HyperloopUPV-H8/Backend-H8/Shared/Server/infra"
-	streaming "github.com/HyperloopUPV-H8/Backend-H8/Shared/StreamingHandlers"
 	excelAdapter "github.com/HyperloopUPV-H8/Backend-H8/Shared/excel_adapter"
 	excelAdapterDomain "github.com/HyperloopUPV-H8/Backend-H8/Shared/excel_adapter/domain"
 	excelRetriever "github.com/HyperloopUPV-H8/Backend-H8/Shared/excel_retriever"
-	transportController "github.com/HyperloopUPV-H8/Backend-H8/Shared/packet_adapter"
+	logger "github.com/HyperloopUPV-H8/Backend-H8/Shared/logger/infra"
+	packetAdapter "github.com/HyperloopUPV-H8/Backend-H8/Shared/packet_adapter"
+	streaming "github.com/HyperloopUPV-H8/Backend-H8/Shared/streaming_handlers"
+	dataTransfer "github.com/HyperloopUPV-H8/Backend-H8/data_transfer"
+	dataTransferDomain "github.com/HyperloopUPV-H8/Backend-H8/data_transfer/domain"
+	messageTransferDomain "github.com/HyperloopUPV-H8/Backend-H8/message_transfer/domain"
+	orderTransferDomain "github.com/HyperloopUPV-H8/Backend-H8/order_transfer/domain"
 	"github.com/joho/godotenv"
 )
 
@@ -29,33 +32,60 @@ func main() {
 		packets = append(packets, board.GetPackets()...)
 	}
 
-	packetAdapter := transportController.New(ips, packets)
-
-	logger := logger.NewLogger(".", time.Second*5)
+	packetAdapter := packetAdapter.New(ips, packets)
 
 	dataTransfer := dataTransfer.New(boards)
+	dataTransfer.Invoke(packetAdapter.ReceiveData)
 
-	server := server.New(dataTransfer.PacketChannel, make(chan any), make(chan any))
+	logger := logger.NewLogger(".", time.Second*5)
+	server := createServer()
 
-	logger.Run()
-
-	server.HandleLog("/backend/log", logger.EnableChan)
-
-	server.HandleWebSocketData("/backend/data", streaming.DataSocketHandler)
-
+	server.HandleLog("/log", logger.EnableChan)
+	server.HandleWebSocketData("/data", streaming.DataSocketHandler)
+	server.HandleWebSocketOrder("/order", streaming.OrderSocketHandler)
+	//server.HandleWebSocketMessage("/message", streaming.MessageSocketHandler)
 	server.HandleSPA()
 
 	go func() {
-		for {
-			payload := dataTransfer.Parse(packetAdapter.ReceiveData)
+		for packetTimestampPair := range dataTransfer.PacketTimestampChannel {
 			select {
-			case logger.EntryChan <- *payload:
+			case logger.EntryChan <- packetTimestampPair:
 			default:
 			}
-			dataTransfer.PacketChannel <- payload.Packet
+			select {
+			case server.PacketChan <- packetTimestampPair.Packet:
+			default:
+			}
+
 		}
 	}()
 
+	//logger.Run()
+
+	// go func() {
+	// 	for {
+	// 		payload := dataTransfer.Parse(packetAdapter.ReceiveData)
+	// 		select {
+	// 		case logger.EntryChan <- *payload:
+	// 		default:
+	// 		}
+	// 		dataTransfer.PacketTimestampChan <- payload.Packet
+	// 	}
+	// }()
+
 	fmt.Println("listening")
 	server.ListenAndServe()
+}
+
+func createServer() server.HTTPServer[
+	dataTransferDomain.Packet,
+	orderTransferDomain.OrderWebAdapter,
+	messageTransferDomain.Message,
+] {
+	server := server.New[
+		dataTransferDomain.Packet,
+		orderTransferDomain.OrderWebAdapter,
+		messageTransferDomain.Message]()
+
+	return server
 }
