+++ src/cmd/MVP-2/static/src/webSocketHandlers.ts
@@ -1,7 +1,7 @@
 import { Packet } from "./modals";
 import { updateReceiveTable } from "./receiveTableBuilder";
 
-let dataSocket = new WebSocket("ws://127.0.0.1:4000/backend/data");
+let dataSocket = new WebSocket("ws://127.0.0.1:4000/data");
 
 dataSocket.onopen = (ev) => {
   alert("Established WS connection");
@@ -13,6 +13,5 @@ dataSocket.onmessage = (ev) => {
   for (let [key, value] of Object.entries(packetsObject)) {
     packetMap.set(Number.parseInt(key), value as Packet);
   }
-  console.log("bruh", packetMap, packetsObject);
   updateReceiveTable(packetMap);
 };
