+++ src/data_transfer/domain/measurement/ranges.go
