+++ src/data_transfer/domain/measurement/value/number/Number.go
