+++ src/data_transfer/domain/measurement/value/string.go
