+++ src/data_transfer/domain/measurement/value/string_test.go
