+++ src/message_transfer/domain/message.go
@@ -1,0 +1,6 @@
+package domain
+
+type Message struct {
+	Id     uint16
+	Fields map[string]string
+}
