+++ src/order_transfer/domain/order_transfer.go
@@ -1,4 +1,4 @@
-package ordertransfer
+package domain
 
 type OrderTransfer struct {
 	OrderWebAdapterChannel chan OrderWebAdapter
